flask_app/requirements.txt
flask==2.3.2
flask-socketio==5.3.4
eventlet==0.33.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==2.2.1
sqlalchemy==2.0.29
eventlet==0.33.3